computer_sensor/requirements.txt
flask
kafka-python
websocket-client
requests